Bump version to 1.0.6, update Kotlin badge, and add Jetpack Compose usage
This commit increments the library version to 1.0.6 and updates the copyright year.
It also updates the Kotlin version badge in the README to 1.9.23.

More significantly, this commit adds a new section to the README.md demonstrating how to use `PDFView` within a Jetpack Compose UI using the `AndroidView` composable. This includes a sample Composable function `PDFViewCompose` and its usage.

Key changes:
- `gradle.properties`:
    - `POM_VERSION` updated from `1.0.5` to `1.0.6`.
    - `POM_CURRENT_YEAR` updated to `2025`.
- `README.md`:
    - Kotlin version badge updated from `1.8.10` to `1.9.23`.
    - Dependency version updated to `1.0.6`.
    - Added "Jetpack Compose Usage" section with a code example.
    - Copyright notice year updated from `2024` to `2023-2025`.